// Src/ShopperHolic.WebUI/ShopperHolic/src/app/models/orderProcessing/returnNotes/returnNoteItem.ts
export class ReturnNoteItem {
    returnNoteItemID: number = 0;
    returnNoteID: number = 0;
    rmaid: number = 0;
    orderID: number = 0;
    returnedDate: Date = new Date();
    orderItemID: number = 0;
    orderItemDescription: string = "";
    returnQty: number = 0;
    customerName: string = "";
    itemCode: string = "";   
}